finish ownership

+++ src/routes/projects/+page.server.ts
@@ -1,9 +1,9 @@
-import type { RequestEvent } from '@sveltejs/kit';
 import { parseRequest } from '@/form-helpers';
 import { projectDeleteSchema, type ProjectDeleteSchemaType, ProjectsService } from '@/server/services/projects.service';
+import type { EnrichedRequestEvent } from '@/types';
 
-export async function load() {
-	const projects = await ProjectsService.findAll();
+export async function load(event: EnrichedRequestEvent) {
+	const projects = await ProjectsService.findAll({userId: event.locals.currentUser.id});
 
 	return {
 		projects
@@ -11,10 +11,10 @@ export async function load() {
 }
 
 export const actions = {
-	delete: async (event: RequestEvent) => {
+	delete: async (event: EnrichedRequestEvent) => {
 		const data = await parseRequest<ProjectDeleteSchemaType>(event, projectDeleteSchema);
 		if (!data) return;
 
-		await ProjectsService.delete(data.id);
+		await ProjectsService.delete(data.id, {userId: event.locals.currentUser.id});
 	}
 };
